[ROC-78] Merge complete. Small refactor. Tests pass.
